Fix browser helper to avoid copying existing crash dumps
Bug: 393986723
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/8018627
Auto-Submit: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Liviu Rau <liviurau@chromium.org>

+++ test/e2e/shared/browser-helper.ts
@@ -36,6 +36,10 @@ export class BrowserWrapper {
       return;
     }
     for (const file of fs.readdirSync(crashesPath)) {
+      const target = path.join(TestConfig.artifactsDir, file);
+      if (fs.existsSync(target)) {
+        continue;
+      }
       console.error('Collecting crash dump:', file);
       fs.copyFileSync(path.join(crashesPath, file), path.join(TestConfig.artifactsDir, file));
     }
